requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
faiss-cpu==1.12.0
idna==3.10
numpy==2.3.2
packaging==25.0
requests==2.32.5
urllib3==2.5.0
